Fix duplicate skill autocasts
This was caused by two different issues. The first was that certain
skills (e.g. regen_hp, regen_mp) have a cooldown timer that is triggered
at the same time as another cooldown timer (e.g. use_hp, use_mp). This
means that while they appear distinct, we need to treat them as one
cooldown.

The second issue is that while unlikely, it is possible for two
different autocasts to be running on the same skill.

The fix for this is to introduce a special token for each autocast.
Whenever a new autocast is created for a skill with the same cooldown we
first use the token to mark the old skill inactive and then replace it
with our token. This way the old autocast will gracefully expire and be
replaced by the new one.