Fix jsconfig
Setting `noLib` results in VSCode not knowing about any built-in types,
even those from the standard HTML DOM. Instead we set `"module": "none"`,
since we don't need to transpile and set `"target": "es2017"` so we have
a bunch of reasonably up-to-date features (e.g. `async`).

Also limit the scope of the `adventureland` submodule include so we
don't pull in the CLI or example code.