Turn Skill and Character into namespaces
Since most functions implicitly use the current character, it's not
terribly useful to have a unique instance of theses classes. Instead
these will be statically initialized at startup.

The exception here is that the `Skill` class is now renamed
`SkillWrapper` and will be exported via the `Skill` namespace which
contains all valid skills for the current class.