Give character a "home" location
The character will return to this location whenever they are idle.
This can be used to keep a character near a group of mobs or provide a
safe location for recovery.

The value is persisted in local storage, so will resume even if the
server reloads. For this reason you will need to explicitly set the home
position using `set_home` to change it.