Merchant: Cast Merchant's luck on nearby players
Don't cast on other merchants or if we've cast it on them recently.
If it's a strong buff, we can't override it since it was cast by someone
in their party.

Fixes a bug where `Skill.wait_until_ready` would never become ready
until the first time it was cast. This is because not all skills are
populated in `parent.next_skill`.