Merchant: Fix exchange logic
The main issue here is that we tried to directly link an item name
to a quest. While some quests have the same name as the item they're
associated with, this may not always be true.

To do this correctly we first build up a mapping of Quest IDs to their
assocated NPC IDs. Then we check `G.items[]` for each exchangeable item
to see what is the assocated quest.

In short: `item_id` → `quest_id` → `npc_id`.