Implement pathfinding in WebWorker
Pathfinding is particularly CPU intensive, so the approach of yielding
the pathfinding function every 10ms was a bit of a hack.

The main challenge here is that in the WebWorker we no longer have
access to any of the common functions, so they need to be reimplemented.

WebWorkers are also picky about their starting origin, so we have to
first create the WebWorker in an anonymous Blob and then import the
worker script dynamically.