Merchant: Smarter bank logic
We now consider both inventory and bank when looking for items to move.

This has a big impact on compounding where we may not have all items we
wish to compound in the bank.

Instead of trying to find empty places, we now look for possible swaps
(e.g. where we want to store one item and withdraw another). This tends
to be much more likely to be successful.